refactor(core): fix typo in get-repl-fn filename
