+++ packages/core/repl/native-functions/get-repl-fn.ts
