+++ packages/core/repl/native-functions/index.ts
@@ -1,5 +1,5 @@
 export * from './help-repl-fn';
-export * from './get-relp-fn';
+export * from './get-repl-fn';
 export * from './resolve-repl-fn';
 export * from './select-relp-fn';
 export * from './debug-repl-fn';
